refactor!: remove CleanButton component
